Full support for local collections
I had to change {$set: updatedData} back to updatedData in the
Constellation_update method (now in common/mutators.js). I notice this
was only for the collection2 update, not for the regular one, which is
odd.  I seem to remember the change was made to fix some edge case, but
can’t remember what it is.  If an issue arises, we can change it back
later, but the local collection undo will break, in that deleted fields
won’t get $unset (actually, the regular collection undo should break
too — I don’t know why it doesn’t!).